requirements.txt
pandas
numpy
openai
sqlalchemy
python-dotenv
pymysql
openpyxl
